Added the UrlBuilder and wrapper for the class.
